pi-agent/requirements.txt
firebase-admin==6.8.0
RPi.GPIO==0.7.1
python-dotenv==1.0.1
